FDG-9704 move data dictionary download to its own file
